Creating subclasses and a superclass Car
