+++ Car.java
@@ -1,0 +1,51 @@
+import java.awt.*;
+
+public abstract class Car {
+    public int nrDoors; // Number of doors on the car
+    public double enginePower; // Engine power of the car
+    public double currentSpeed; // The current speed of the car
+    public Color color; // Color of the car
+    public String modelName; // The car model name
+
+    public int getNrDoors(){
+        return nrDoors;
+    }
+    public double getEnginePower(){
+        return enginePower;
+    }
+
+    public double getCurrentSpeed(){
+        return currentSpeed;
+    }
+
+    public Color getColor(){
+        return color;
+    }
+
+    public void setColor(Color clr){
+	    color = clr;
+    }
+
+    public void startEngine(){
+	    currentSpeed = 0.1;
+    }
+
+    public void stopEngine(){
+	    currentSpeed = 0;
+    }
+    
+    // TODO fix this method according to lab pm
+    public void gas(double amount){
+        incrementSpeed(amount);
+    }
+
+    // TODO fix this method according to lab pm
+    public void brake(double amount){
+        decrementSpeed(amount);
+    }
+    
+    abstract void incrementSpeed(double amount9);
+
+    abstract void decrementSpeed(double amount);
+
+}
