+++ Volvo240.java
@@ -1,6 +1,6 @@
 import java.awt.*;
 
-public class Volvo240{
+public class Volvo240 extends Car{
 
     public final static double trimFactor = 1.25;
     public int nrDoors; // Number of doors on the car
@@ -17,32 +17,6 @@ public class Volvo240{
         stopEngine();
     }
     
-    public int getNrDoors(){
-        return nrDoors;
-    }
-    public double getEnginePower(){
-        return enginePower;
-    }
-
-    public double getCurrentSpeed(){
-        return currentSpeed;
-    }
-
-    public Color getColor(){
-        return color;
-    }
-
-    public void setColor(Color clr){
-	    color = clr;
-    }
-
-    public void startEngine(){
-	    currentSpeed = 0.1;
-    }
-
-    public void stopEngine(){
-	    currentSpeed = 0;
-    }
     
     public double speedFactor(){
         return enginePower * 0.01 * trimFactor;
@@ -56,13 +30,4 @@ public class Volvo240{
         currentSpeed = Math.max(getCurrentSpeed() - speedFactor() * amount,0);
     }
 
-    // TODO fix this method according to lab pm
-    public void gas(double amount){
-        incrementSpeed(amount);
-    }
-
-    // TODO fix this method according to lab pm
-    public void brake(double amount){
-        decrementSpeed(amount);
-    }
 }
